+++ synergyZone/.settings/org.eclipse.core.resources.prefs
@@ -4,6 +4,6 @@
 eclipse.preferences.version=1
 encoding//src/main/java=UTF-8
 encoding//src/main/resources=UTF-8
-encoding//src/main/webapp/WEB-INF/views/employee/detail.jsp=UTF-8
+encoding//src/main/webapp/WEB-INF/views/admin/detail.jsp=UTF-8
 encoding//src/test/java=UTF-8
 encoding/<project>=UTF-8
