--- synergyZone/src/main/java/com/kh/synergyZone/component/EmpNoGenerator.java
@@ -1,6 +1,0 @@
-package com.kh.synergyZone.component;
-
-public interface EmpNoGenerator {
-	String generateEmpNo(String deptNo);
-	
-}
