+++ synergyZone/src/main/java/com/kh/synergyZone/controller/AddressController.java
@@ -63,9 +63,6 @@ public class AddressController {
 
        return "address/list";
    }
-   
-    
-   
     
     
 }
