+++ synergyZone/src/main/webapp/WEB-INF/views/admin/detail.jsp
