+++ synergyZone/src/main/webapp/WEB-INF/views/admin/edit.jsp
