+++ synergyZone/src/main/webapp/WEB-INF/views/admin/job/list.jsp
