+++ synergyZone/src/main/webapp/WEB-INF/views/admin/job/register.jsp
