+++ synergyZone/src/main/webapp/WEB-INF/views/admin/join.jsp
