+++ synergyZone/src/main/webapp/WEB-INF/views/admin/list.jsp
