+++ synergyZone/src/main/webapp/WEB-INF/views/admin/waitingList.jsp
